refactor: extract filtering code
